# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: MI Academy API
  version: "1.0.0"
  description: |
    MI Academy backend API — parent dashboard, child progress, games, and sync.
    All contracts are versioned. Breaking changes increment the major version.
  contact:
    name: MI Academy Platform Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Current production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AgeGroup:
      type: string
      enum: [junior, explorer, master]
    AccessibilityPreferencesSchema:
      type: object
      properties:
        high_contrast:
          type: boolean
          default: false
        large_text:
          type: boolean
          default: false
        reduce_motion:
          type: boolean
          default: false
        screen_reader:
          type: boolean
          default: false
        font_size:
          type: number
          minimum: 0.8
          maximum: 2.0
          default: 1.0
    AudioPreferencesSchema:
      type: object
      properties:
        music_volume:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.8
        sfx_volume:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 1.0
        speech_enabled:
          type: boolean
          default: true
...
    GameListItem:
      type: object
      required: [id, name, game_type, is_active]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        game_type:
          type: string
          enum: [memory, word_builder, sound_match, math_race, math_supermarket, robot_commands]
        age_min:
          type: integer
        age_max:
          type: integer
        is_active:
          type: boolean
    GameDetail:
      type: object
      required: [id, name, game_type, is_active]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        game_type:
          type: string
          enum: [memory, word_builder, sound_match, math_race, math_supermarket, robot_commands]
        age_min:
          type: integer
        age_max:
          type: integer
        config_json:
          type: object
          additionalProperties: true
        is_active:
          type: boolean
    GameStartRequest:
      type: object
      required: [child_id]
      properties:
        child_id:
          type: string
          format: uuid
    GameAttemptRequest:
      type: object
      required: [child_id, answer_json, response_time_ms, hint_count]
      properties:
        child_id:
          type: string
          format: uuid
        answer_json:
          type: object
          additionalProperties: true
        response_time_ms:
          type: integer
        hint_count:
          type: integer
          minimum: 0
    GameCompleteRequest:
      type: object
      required: [child_id, total_stars]
      properties:
        child_id:
          type: string
          format: uuid
        total_stars:
          type: integer
          minimum: 0
          maximum: 3
        badges_unlocked:
          type: array
          items:
            type: string
          default: []
        rewards_unlocked:
          type: array
          items:
            type: string
          default: []
    MiGameLaunchRequest:
      type: object
      description: Full game launch request from platform to game. NEVER contains tokens or PII.
      required: [schema_version, child_profile_id, game_id, level_id, language, age_group, accessibility, audio_preferences, level_content]
      properties:
        schema_version:
          type: integer
          default: 1
        child_profile_id:
          type: string
          format: uuid
        game_id:
          type: string
          format: uuid
        level_id:
          type: string
          format: uuid
        language:
          type: string
          enum: [vi, en]
        age_group:
          $ref: "#/components/schemas/AgeGroup"
        accessibility:
          $ref: "#/components/schemas/AccessibilityPreferencesSchema"
        audio_preferences:
          $ref: "#/components/schemas/AudioPreferencesSchema"
        level_content:
          type: object
          additionalProperties: true
        restored_state:
          type: object
          additionalProperties: true
...
paths:
...
  /games:
    get:
      tags: [games]
      summary: List games
      parameters:
        - name: age_group
          in: query
          schema:
            $ref: "#/components/schemas/AgeGroup"
      responses:
        "200":
          description: Game list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameListItem"
  /games/{game_id}:
    get:
      tags: [games]
      summary: Get game details
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
  /games/{game_id}/start:
    post:
      tags: [games]
      summary: Start game session
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameStartRequest"
      responses:
        "200":
          description: Game session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiGameLaunchRequest"
  /games/{game_id}/attempt:
    post:
      tags: [games]
      summary: Submit game attempt
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameAttemptRequest"
      responses:
        "200":
          description: Attempt recorded
  /games/{game_id}/complete:
    post:
      tags: [games]
      summary: Complete game session
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCompleteRequest"
      responses:
        "200":
          description: Game completed
...